refactor: audit and policy commands to single files
