+++ src/gmp/commands/__tests__/audit.test.ts
@@ -1,18 +1,15 @@
-/* SPDX-FileCopyrightText: 2024 Greenbone AG
+/* SPDX-FileCopyrightText: 2026 Greenbone AG
  *
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
 import {describe, test, expect} from '@gsa/testing';
 import AuditCommand from 'gmp/commands/audit';
-import AuditsCommand from 'gmp/commands/audits';
 import {
   createActionResultResponse,
   createEntityResponse,
-  createEntitiesResponse,
   createHttp,
 } from 'gmp/commands/testing';
-import {ALL_FILTER} from 'gmp/models/filter';
 import {
   OPENVAS_SCANNER_TYPE,
   OPENVAS_DEFAULT_SCANNER_ID,
@@ -244,56 +241,5 @@ describe('AuditCommand tests', () => {
     });
     const {data} = resp;
     expect(data.id).toEqual('foo');
-  });
-});
-
-describe('AuditsCommand tests', () => {
-  test('should return all audits', async () => {
-    const response = createEntitiesResponse('task', [
-      {
-        _id: '1',
-      },
-      {
-        _id: '2',
-      },
-    ]);
-
-    const fakeHttp = createHttp(response);
-
-    const cmd = new AuditsCommand(fakeHttp);
-    const resp = await cmd.getAll();
-    expect(fakeHttp.request).toHaveBeenCalledWith('get', {
-      args: {
-        cmd: 'get_tasks',
-        filter: ALL_FILTER.toFilterString(),
-        usage_type: 'audit',
-      },
-    });
-    const {data} = resp;
-    expect(data.length).toEqual(2);
-  });
-
-  test('should return audits', async () => {
-    const response = createEntitiesResponse('task', [
-      {
-        _id: '1',
-      },
-      {
-        _id: '2',
-      },
-    ]);
-
-    const fakeHttp = createHttp(response);
-
-    const cmd = new AuditsCommand(fakeHttp);
-    const resp = await cmd.get();
-    expect(fakeHttp.request).toHaveBeenCalledWith('get', {
-      args: {
-        cmd: 'get_tasks',
-        usage_type: 'audit',
-      },
-    });
-    const {data} = resp;
-    expect(data.length).toEqual(2);
   });
 });
