+++ src/gmp/commands/__tests__/audits.test.ts
@@ -1,16 +1,16 @@
-/* SPDX-FileCopyrightText: 2024 Greenbone AG
+/* SPDX-FileCopyrightText: 2026 Greenbone AG
  *
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
 import {describe, test, expect} from '@gsa/testing';
-import PoliciesCommand from 'gmp/commands/policies';
+import AuditsCommand from 'gmp/commands/audits';
 import {createEntitiesResponse, createHttp} from 'gmp/commands/testing';
 import {ALL_FILTER} from 'gmp/models/filter';
 
-describe('PoliciesCommand tests', () => {
-  test('should return all policies', async () => {
-    const response = createEntitiesResponse('config', [
+describe('AuditsCommand tests', () => {
+  test('should return all audits', async () => {
+    const response = createEntitiesResponse('task', [
       {
         _id: '1',
       },
@@ -20,21 +20,22 @@ describe('PoliciesCommand tests', () => {
     ]);
 
     const fakeHttp = createHttp(response);
-    const cmd = new PoliciesCommand(fakeHttp);
+
+    const cmd = new AuditsCommand(fakeHttp);
     const resp = await cmd.getAll();
     expect(fakeHttp.request).toHaveBeenCalledWith('get', {
       args: {
-        cmd: 'get_configs',
+        cmd: 'get_tasks',
         filter: ALL_FILTER.toFilterString(),
-        usage_type: 'policy',
+        usage_type: 'audit',
       },
     });
     const {data} = resp;
     expect(data.length).toEqual(2);
   });
 
-  test('should return policies', async () => {
-    const response = createEntitiesResponse('config', [
+  test('should return audits', async () => {
+    const response = createEntitiesResponse('task', [
       {
         _id: '1',
       },
@@ -44,12 +45,13 @@ describe('PoliciesCommand tests', () => {
     ]);
 
     const fakeHttp = createHttp(response);
-    const cmd = new PoliciesCommand(fakeHttp);
+
+    const cmd = new AuditsCommand(fakeHttp);
     const resp = await cmd.get();
     expect(fakeHttp.request).toHaveBeenCalledWith('get', {
       args: {
-        cmd: 'get_configs',
-        usage_type: 'policy',
+        cmd: 'get_tasks',
+        usage_type: 'audit',
       },
     });
     const {data} = resp;
