+++ src/gmp/gmp.ts
@@ -30,7 +30,8 @@ import AgentInstallersCommand from 'gmp/commands/agent-installers';
 import AgentsCommand from 'gmp/commands/agents';
 import AlertCommand from 'gmp/commands/alert';
 import AlertsCommand from 'gmp/commands/alerts';
-import AuditCommand, {AuditsCommand} from 'gmp/commands/audits';
+import AuditCommand from 'gmp/commands/audit';
+import AuditsCommand from 'gmp/commands/audits';
 import AuthenticationCommand from 'gmp/commands/auth';
 import CertBundAdvisoriesCommand from 'gmp/commands/cert-bund-advisories';
 import CertBundAdvisoryCommand from 'gmp/commands/cert-bund-advisory';
@@ -55,7 +56,8 @@ import OciImageTargetsCommand from 'gmp/commands/oci-image-targets';
 import PerformanceCommand from 'gmp/commands/performance';
 import PermissionCommand from 'gmp/commands/permission';
 import PermissionsCommand from 'gmp/commands/permissions';
-import PolicyCommand, {PoliciesCommand} from 'gmp/commands/policies';
+import PoliciesCommand from 'gmp/commands/policies';
+import PolicyCommand from 'gmp/commands/policy';
 import {PortListCommand, PortListsCommand} from 'gmp/commands/port-lists';
 import ReportCommand from 'gmp/commands/report';
 import ReportsCommand from 'gmp/commands/reports';
